// src/pages/MainForms/utils/types/ChildhoodInfoInputs.ts
export type ChildhoodInfoInputs = {
  formId: number
  patientId: number
  frequentColds: boolean
  hasAllergies: boolean
  chronicIllnesses: string
  headSupportAge: string
  crawlingAge: string
  sittingAge: string
  standingAge: string
  walkingAge: string
  climbingAge: string
  towerBuildingAge: string
  monosyllabicWordsAge: string
  twoWordSentencesAge: string
  learnedAndForgottenSkills: string
}